requirements.txt
discord.py
py-cord
aiosqlite
python-dotenv